+++ Flipper/src/Elements/Obstacle.java
@@ -5,11 +5,11 @@ public class Obstacle extends ElementStatic{
         
     }
 
-    public Obstacle(float nx, float ny){
+    public Obstacle(double nx, double ny){
         this.nX=nx;
         this.nY=ny;
     }
-    public Obstacle(float nx, float ny, float s, float e){
+    public Obstacle(double nx, double ny, double s, double e){
         this(nx,ny);
         this.smoothness = s;
         this.elasticity = e;
